Decouple container styles/state
